refactor: move toMovieEntity mapper to data layer
